apply form permission policy to all page controller actions
For the time being, we only want to allows user to create/edit/delete
pages if the pages form belongs to the users organisation

+++ app/controllers/forms/delete_confirmation_controller.rb
@@ -78,6 +78,7 @@ module Forms
     def delete_page(form, page)
       success_url = form_path(form)
 
+      authorize form, :can_view_form?
       if page.destroy
         flash[:message] = "Successfully deleted #{page.question_text}"
         redirect_to success_url, status: :see_other
